final src

+++ src/java/controller/StaffUpdateProductController.java
@@ -24,7 +24,7 @@ import javax.servlet.http.HttpSession;
 
 /**
  *
- * @author THUAN
+ * @author VU
  */
 public class StaffUpdateProductController extends HttpServlet {
 private final String MANGER_PAGE = "StaffUpdateProduct.jsp";
@@ -117,7 +117,7 @@ private final String MANGER_P = "StaffController";
             int PublishDate = Integer.parseInt(request.getParameter("PublishDate"));
             int Quantity = Integer.parseInt(request.getParameter("Quantity"));
             int SubCategoryId = Integer.parseInt(request.getParameter("SubCategoryId"));
-            float UnitPrice = Float.parseFloat(request.getParameter("UnitPrice"));
+            int UnitPrice = Integer.parseInt(request.getParameter("UnitPrice"));
             int CategoryID = Integer.parseInt(request.getParameter("CategoryID"));
             int Status = Integer.parseInt(request.getParameter("Status"));
             int TotalFeedback = Integer.parseInt(request.getParameter("TotalFeedback"));
